Avatar Uploading
* Added Github url

* :stuck_out_tongue: Fixed portfolio controller to redirect to /auth

when not logged in. Shouldn't matter much anyways as the menu won't show the option unless the user is logged in. Maybe this should be fixed?

* :sweat_smile: Added players page to display all the users.

* :relaxed: Added player search

Fixed menu layout as well.

* :wink: Fix from assignment 1: TODO Update

* :sweat: Read description for detail

Added bootstrap
Added updated font-awesome files
Added images to links
Fixed registration authentication

* Started admin console. Forgot to merge

* :skull: Going to work on the console later

* :zzz: Added edit/delete to admin console. Edit not yet working. Accidentally added the changed files to the old commit when fixing

* :fire: Fixed admin new user issues

Added users to the database for the original users.

* :sunglasses: Editing players works. Admin console complete.

Update todo and changelog

* :relaxed: Added user avatar uploading. Can also change avatar after uploading. Added default image.